+++ scenes/ui/menus/pause_menu.gd
@@ -1,23 +1,23 @@
-extends Control
+extends Menu
 
-signal game_restarted
-
-@export var default_focus: Control
+signal menu_pressed(previous: Menu)
+signal options_pressed(previous: Menu)
 
 var can_pause := false
 
 
 func set_paused(paused: bool) -> void:
-	if not self.can_pause:
+	if not self.can_pause or paused == self.visible:
 		return
 	self.get_tree().paused = paused
-	self.visible = paused
 	if paused:
-		self.default_focus.grab_focus()
+		self.open()
+	else:
+		self.close()
 
 
-func _unhandled_input(event: InputEvent) -> void:
-	if event.is_action_released("pause"):
+func _input(event: InputEvent) -> void:
+	if event.is_action_released(&"pause"):
 		self.set_paused(not self.get_tree().paused)
 
 
@@ -28,10 +28,10 @@ func _on_resume_button_pressed() -> void:
 func _on_menu_button_pressed() -> void:
 	self.set_paused(false)
 	self.can_pause = false
-	self.game_restarted.emit()
+	self.menu_pressed.emit(self)
 
 
-func _on_lobby_menu_players_ready() -> void:
+func _on_lobby_menu_play_pressed(_previous: Menu = null) -> void:
 	self.can_pause = true
 
 
@@ -40,5 +40,10 @@ func _on_main_game_over() -> void:
 
 
 func _notification(what: int) -> void:
-	if what == NOTIFICATION_WM_WINDOW_FOCUS_OUT and self.can_pause:
+	if what == NOTIFICATION_WM_WINDOW_FOCUS_OUT:
 		self.set_paused(true)
+
+
+func _on_options_button_pressed() -> void:
+	self.hide()
+	self.options_pressed.emit(self)