+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow setting max upgrade points in the lobby
 - Scale ship rotation speed by controller axis strength
 - Pause when window loses focus
+- Allow accessing the options menu from the pause menu
 - Stop timer after game ends in Last Man Standing mode
 - Tiebreak by percent health remaining, not total health remaining